Temporary Sync Method
Hoh boy. So /b/ goes down for an hour on the dot, and when it comes
back, no XHRs work to 2-hi.me. Struggled with that for hours, I can see
post requests leaving with all data in tact, yet when the arrive at the
server, nada. Tried this with $.ajax and GM_xmlHttpRequest with many
methods and variations, all to no luck. So, temporarily, we're going
stoneage, and using a hidden ifame to submit the posts to 2-hi.me.
Silver lining: image uploads work, they couldn't be done with XHRs.
Still don't show images yet tho :3